[Refactor] #139 변경된 ddd 내용에 맞게 dto 디렉토리 위치 수정
