+++ performance-service/src/main/java/com/ticketrush/boundedcontext/performance/domain/entity/Performance.java
@@ -1,6 +1,6 @@
 package com.ticketrush.boundedcontext.performance.domain.entity;
 
-import com.ticketrush.boundedcontext.performance.domain.dto.request.PerformanceCreateRequest;
+import com.ticketrush.boundedcontext.performance.app.dto.request.PerformanceCreateRequest;
 import com.ticketrush.boundedcontext.performance.domain.types.Genre;
 import com.ticketrush.boundedcontext.performance.domain.types.PerformanceStatus;
 import com.ticketrush.global.jpa.entity.AutoIdBaseEntity;
