+++ seat-service/src/main/java/com/ticketrush/boundedcontext/seat/app/dto/response/SeatLayoutResponse.java
@@ -1,3 +1,3 @@
-package com.ticketrush.boundedcontext.seat.domain.dto.response;
+package com.ticketrush.boundedcontext.seat.app.dto.response;
 
 public record SeatLayoutResponse(Long seatId, Long seatLayoutId, String rowNo, Integer colNo) {}
