+++ seat-service/src/main/java/com/ticketrush/boundedcontext/seat/app/facade/SeatFacade.java
@@ -1,7 +1,7 @@
 package com.ticketrush.boundedcontext.seat.app.facade;
 
+import com.ticketrush.boundedcontext.seat.app.dto.response.SeatLayoutResponse;
 import com.ticketrush.boundedcontext.seat.app.usecase.SeatGetSeatLayoutsUseCase;
-import com.ticketrush.boundedcontext.seat.domain.dto.response.SeatLayoutResponse;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
